refactor(agents): extract candidate adoption from cook.rs
cook.rs regrew from ~2705 to 3665 LOC since the earlier decomposition (#9120/
#9133/#9137) — ~960 LOC of retry/adoption-gate/recovery fixes (#9149/#9209/#9154/
#9212/#9172/#9193/#9231/#9192) accreted onto the run_cook spine instead of a
dedicated module. This carves the largest cohesive cluster back out.

Extracts the candidate-adoption family into cook_adoption.rs: adopt_cook_candidate
(+_with_dispatcher/_with_options_and_dispatcher/_with_dispatcher_and_backend), plus
the adoption resolution helpers (resolve_cook_adoption_attempt/resolve_adoption_target/
candidate_adoption_source/concrete_adoption_ai_model) and gate-failure comparison
(compare_adoption_gate_failures_to_base). Adoption admits an externally prepared
immutable commit into a durable cook without replaying provider work — a cohesive
boundary distinct from the run_cook execution engine.

Behavior-preserving pure move following the agent_task_service module pattern
(mod X; pub use X::*). New module has explicit imports (matching cook_promotion/
cook_baseline siblings) and references cook.rs-owned types via super::cook::.
Widened gate_feedback_current_diff to pub(crate) for the run_cook caller, and the
four private adoption helpers to pub(crate) for the cook.rs test module. Trimmed
cook.rs imports that moved with the cluster.

cook.rs: 3665 -> 3037 LOC.
