+++ crates/homeboy-agents/src/agent_task_service/cook.rs
@@ -6,25 +6,15 @@ use serde_json::Value;
 use std::path::PathBuf;
 use std::sync::{mpsc, Arc, Mutex};
 
-use crate::agent_task::{AgentTaskExecutor, AgentTaskRequest};
+use crate::agent_task::AgentTaskExecutor;
 use crate::agent_task_cook_loop::{
     evaluate_cook_loop, AgentTaskCookLoopOptions, AgentTaskCookLoopReport, AgentTaskCookLoopStatus,
 };
 use crate::agent_task_dispatch_plan::build_dispatch_plan;
 use crate::agent_task_dispatch_service::{self, AgentTaskDispatchCommand};
-use crate::agent_task_finalization::{
-    AgentTaskPrFinalizationBackend, RealAgentTaskPrFinalizationBackend,
-};
 use crate::agent_task_gate::VerifyGateOptions;
-use crate::agent_task_gate::{
-    failure_fingerprint, run_gate_command_with_timeout, AgentTaskGateBaselineComparison,
-    AgentTaskGateStatus,
-};
 use crate::agent_task_lifecycle;
-use crate::agent_task_promotion::resolve_candidate_revision;
-use crate::agent_task_promotion::{
-    promote_with_checkpoint, AgentTaskPromotionOptions, AgentTaskPromotionReport,
-};
+use crate::agent_task_promotion::AgentTaskPromotionReport;
 use crate::agent_task_scheduler::{
     AgentTaskExecutionBudget, AgentTaskExecutorAdapter, AgentTaskPlan,
 };
@@ -39,16 +29,14 @@ use super::cook_budget::{
     budget_remaining, execution_budget_usage, reserve_remediation_budget, ExecutionBudgetUsage,
 };
 use super::cook_promotion::{
-    attempt_needs_execution, cook_report, finalize_or_load_cook_pr,
-    finalize_or_load_cook_pr_with_backend, persisted_promotion_for_attempt,
-    promote_or_load_attempt, promotion_source,
+    attempt_needs_execution, cook_report, finalize_or_load_cook_pr, promote_or_load_attempt,
 };
 use super::execution::run_loaded_plan_with_derived_cook_baseline;
 use super::AgentTaskRunResult;
 
 /// The promotion checkpoint captures this before gates run, when it is the only
 /// complete authorization for reusing the dirty managed destination.
-fn gate_feedback_current_diff(promotion: &AgentTaskPromotionReport) -> String {
+pub(crate) fn gate_feedback_current_diff(promotion: &AgentTaskPromotionReport) -> String {
     promotion
         .provenance
         .pointer("/gate_feedback_baseline/current_diff")
@@ -145,634 +133,6 @@ pub struct AgentTaskCookReport {
     pub terminal_phase: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub terminal_failure_classification: Option<String>,
-}
-
-/// Adopt an externally prepared immutable commit into a durable cook. The
-/// source recipe remains authoritative for repository, base, gates, and
-/// finalization policy; adoption only replaces provider artifact harvesting.
-pub fn adopt_cook_candidate(
-    cook_or_run_id: &str,
-    candidate_ref: &str,
-) -> Result<AgentTaskRunResult<AgentTaskCookReport>> {
-    adopt_cook_candidate_with_options_and_dispatcher(
-        cook_or_run_id,
-        candidate_ref,
-        AgentTaskCandidateAdoptionOptions::default(),
-        |_| Ok(None),
-    )
-}
-
-/// Compatibility entry point for callers that previously supplied attempt
-/// transport reconstruction. Candidate adoption never replays provider work,
-/// so the dispatcher is intentionally not reconstructed or prepared.
-pub fn adopt_cook_candidate_with_dispatcher(
-    cook_or_run_id: &str,
-    candidate_ref: &str,
-    reconstruct_dispatcher: impl FnOnce(
-        &Value,
-    ) -> Result<Option<Arc<dyn AgentTaskCookAttemptDispatcher>>>,
-) -> Result<AgentTaskRunResult<AgentTaskCookReport>> {
-    adopt_cook_candidate_with_options_and_dispatcher(
-        cook_or_run_id,
-        candidate_ref,
-        AgentTaskCandidateAdoptionOptions::default(),
-        reconstruct_dispatcher,
-    )
-}
-
-/// Adopt a candidate with provenance supplied by the external preparer.
-pub fn adopt_cook_candidate_with_options_and_dispatcher(
-    cook_or_run_id: &str,
-    candidate_ref: &str,
-    adoption: AgentTaskCandidateAdoptionOptions,
-    reconstruct_dispatcher: impl FnOnce(
-        &Value,
-    ) -> Result<Option<Arc<dyn AgentTaskCookAttemptDispatcher>>>,
-) -> Result<AgentTaskRunResult<AgentTaskCookReport>> {
-    adopt_cook_candidate_with_dispatcher_and_backend(
-        cook_or_run_id,
-        candidate_ref,
-        adoption,
-        reconstruct_dispatcher,
-        &mut RealAgentTaskPrFinalizationBackend,
-    )
-}
-
-fn adopt_cook_candidate_with_dispatcher_and_backend<B: AgentTaskPrFinalizationBackend>(
-    cook_or_run_id: &str,
-    candidate_ref: &str,
-    adoption: AgentTaskCandidateAdoptionOptions,
-    _reconstruct_dispatcher: impl FnOnce(
-        &Value,
-    ) -> Result<Option<Arc<dyn AgentTaskCookAttemptDispatcher>>>,
-    backend: &mut B,
-) -> Result<AgentTaskRunResult<AgentTaskCookReport>> {
-    let (record, recipe) = resolve_adoption_target(cook_or_run_id)?;
-    let cook_id = &recipe.cook_id;
-    let mut options = super::reconstruct_adoption_options(&recipe)?;
-    let run_id = record.run_id.clone();
-    let plan = agent_task_lifecycle::load_plan(&run_id)?;
-    let source_request = plan.tasks.first().cloned().ok_or_else(|| {
-        Error::validation_invalid_argument(
-            "run_id",
-            "candidate adoption requires a cook run with one source task",
-            Some(run_id.clone()),
-            None,
-        )
-    })?;
-    if plan.tasks.len() != 1 {
-        return Err(Error::validation_invalid_argument(
-            "run_id",
-            "candidate adoption supports one task per cook",
-            Some(run_id),
-            None,
-        ));
-    }
-    let (source, source_path, recovery) = candidate_adoption_source(&record, &source_request)?;
-    let (adoption_ai_model, ai_model_source) = match adoption.ai_model {
-        Some(model) => (concrete_adoption_ai_model(&model)?, "candidate_input"),
-        None => (
-            concrete_adoption_ai_model(options.ai_model.as_deref().unwrap_or_default())?,
-            "recipe_finalization",
-        ),
-    };
-    let source_worktree = options.source_worktree_path.as_deref().ok_or_else(|| {
-        Error::validation_invalid_argument(
-            "candidate_ref",
-            "candidate adoption requires the recorded source worktree",
-            None,
-            None,
-        )
-    })?;
-    // Resolve the caller input to the commit object before durable ownership is
-    // claimed, then use that immutable SHA for every subsequent operation.
-    let candidate_sha = resolve_candidate_revision(source_worktree, candidate_ref)?;
-    let gate_identity = if options.gates.verify.is_empty() {
-        "promotion verification".to_string()
-    } else {
-        options.gates.verify.join(" && ")
-    };
-    if !options.gates.rerun_completed_gates
-        && record.candidate_adoption.as_ref().is_some_and(|adoption| {
-            adoption.state == "completed"
-                && adoption.candidate_sha == candidate_sha
-                && adoption.ai_model == adoption_ai_model
-        })
-    {
-        let promotion = persisted_promotion_for_attempt(&record.run_id)?.ok_or_else(|| {
-            Error::validation_invalid_argument(
-                "candidate_ref",
-                "completed candidate adoption is missing its persisted promotion result",
-                Some(candidate_sha.clone()),
-                None,
-            )
-        })?;
-        let feedback = evaluate_cook_loop(AgentTaskCookLoopOptions {
-            source_request,
-            promotion_report: promotion.clone(),
-            attempt: 1,
-            max_attempts: options.max_attempts,
-            source_run_id: Some(record.run_id.clone()),
-            current_diff: String::new(),
-            metadata: serde_json::json!({"adopted_candidate_ref": candidate_ref}),
-        });
-        let finalization = record.metadata.get("cook_finalization").cloned();
-        let status = finalization
-            .as_ref()
-            .and_then(|value| value["status"].as_str())
-            .unwrap_or("green_no_finalize")
-            .to_string();
-        return Ok(cook_report(
-            cook_id.to_string(),
-            &status,
-            vec![AgentTaskCookAttemptReport {
-                attempt: 1,
-                run_id: record.run_id.clone(),
-                run_state: format!("{:?}", record.state),
-                aggregate_path: record.aggregate_path.clone(),
-                promotion: Some(promotion),
-                feedback: Some(feedback),
-            }],
-            finalization,
-            Some("reused the completed candidate adoption result; set rerun_completed_gates to rerun its gates".to_string()),
-            if status == "review_ready" || status == "green_no_finalize" { 0 } else { 1 },
-        ));
-    }
-    agent_task_lifecycle::start_candidate_adoption_with_rerun_policy(
-        &record.run_id,
-        &candidate_sha,
-        &adoption_ai_model,
-        &gate_identity,
-        options.gates.rerun_completed_gates,
-    )?;
-    let gate_run_id = record.run_id.clone();
-    let promotion = crate::agent_task_promotion::with_gate_supervision(
-        crate::agent_task_gate::GateSupervision {
-            timeout: options.gates.gate_timeout(),
-            heartbeat_interval: options.gates.gate_heartbeat_interval(),
-            on_spawn: Arc::new({
-                let run_id = gate_run_id.clone();
-                move |pid, command| {
-                    agent_task_lifecycle::start_candidate_adoption_gate(
-                        &run_id,
-                        command,
-                        pid,
-                        options.gates.gate_timeout_seconds,
-                    )
-                }
-            }),
-            on_heartbeat: Arc::new({
-                let run_id = gate_run_id.clone();
-                move |tail| agent_task_lifecycle::heartbeat_candidate_adoption_gate(&run_id, tail)
-            }),
-            is_cancelled: Arc::new(move || {
-                agent_task_lifecycle::candidate_adoption_cancel_requested(&gate_run_id)
-                    .unwrap_or(false)
-            }),
-        },
-        || {
-            promote_with_checkpoint(
-                AgentTaskPromotionOptions {
-                    source,
-                    source_run_id: Some(record.run_id.clone()),
-                    source_path,
-                    source_worktree_path: options.source_worktree_path.clone(),
-                    base_ref: Some(options.base.clone()),
-                    task_base_sha: options.task_base_sha.clone(),
-                    candidate_ref: Some(candidate_sha.clone()),
-                    to_worktree: options.to_worktree.clone(),
-                    task_id: None,
-                    artifact_id: None,
-                    dry_run: false,
-                    gates: options.gates.clone(),
-                    provider_command: options.provider_command.clone(),
-                    provider_invocation: options.provider_invocation.clone(),
-                },
-                |checkpoint| {
-                    let checkpoint = serde_json::to_value(checkpoint).map_err(|error| {
-                        Error::internal_json(
-                            error.to_string(),
-                            Some("serialize adopted candidate checkpoint".to_string()),
-                        )
-                    })?;
-                    agent_task_lifecycle::checkpoint_candidate_adoption(
-                        &record.run_id,
-                        "post_apply_verification",
-                        &gate_identity,
-                    )?;
-                    agent_task_lifecycle::record_promotion(&record.run_id, checkpoint).map(|_| ())
-                },
-            )
-        },
-    );
-    let mut promotion = match promotion {
-        Ok(promotion) => promotion,
-        Err(error) => {
-            agent_task_lifecycle::finish_candidate_adoption(
-                &record.run_id,
-                Some(error.message.clone()),
-            )?;
-            return Err(error);
-        }
-    };
-    if agent_task_lifecycle::candidate_adoption_cancel_requested(&record.run_id)? {
-        return Err(Error::validation_invalid_argument(
-            "run_id",
-            "candidate adoption was cancelled before baseline verification",
-            Some(record.run_id.clone()),
-            None,
-        ));
-    }
-    let task_base_sha = options.task_base_sha.as_deref().ok_or_else(|| {
-        Error::validation_invalid_argument(
-            "task_base_sha",
-            "candidate adoption requires the recorded immutable task base for baseline-aware verification",
-            None,
-            None,
-        )
-    })?;
-    compare_adoption_gate_failures_to_base(
-        &mut promotion,
-        source_worktree,
-        task_base_sha,
-        &record.run_id,
-    )?;
-    if agent_task_lifecycle::candidate_adoption_cancel_requested(&record.run_id)? {
-        return Err(Error::validation_invalid_argument(
-            "run_id",
-            "candidate adoption was cancelled before promotion could finalize",
-            Some(record.run_id.clone()),
-            None,
-        ));
-    }
-    // The adopted candidate did not run through this cook's provider lifecycle.
-    // Bind its declared model to the authenticated promotion instead of inferring
-    // one from the immutable execution plan.
-    options.ai_model = Some(adoption_ai_model.clone());
-    promotion.provenance["adoption"] = serde_json::json!({
-        "source_run_id": record.run_id,
-        "candidate_ref": candidate_sha,
-        "source_worktree_path": options.source_worktree_path,
-        "recorded_task_base": options.task_base_sha,
-        "recovery": recovery,
-        "ai_model": adoption_ai_model,
-        "ai_model_source": ai_model_source,
-    });
-    let promotion_value = serde_json::to_value(&promotion)
-        .map_err(|error| Error::internal_json(error.to_string(), None))?;
-    agent_task_lifecycle::record_promotion(&record.run_id, promotion_value)?;
-    let feedback = evaluate_cook_loop(AgentTaskCookLoopOptions {
-        source_request,
-        promotion_report: promotion.clone(),
-        attempt: 1,
-        max_attempts: options.max_attempts,
-        source_run_id: Some(record.run_id.clone()),
-        current_diff: gate_feedback_current_diff(&promotion),
-        metadata: serde_json::json!({"adopted_candidate_ref": candidate_ref}),
-    });
-    let attempt = AgentTaskCookAttemptReport {
-        attempt: 1,
-        run_id: record.run_id.clone(),
-        run_state: format!("{:?}", record.state),
-        aggregate_path: record.aggregate_path.clone(),
-        promotion: Some(promotion.clone()),
-        feedback: Some(feedback.clone()),
-    };
-    if feedback.status != AgentTaskCookLoopStatus::GreenCompleted {
-        agent_task_lifecycle::finish_candidate_adoption(
-            &record.run_id,
-            Some("adopted candidate did not pass the original deterministic gates".to_string()),
-        )?;
-        return Ok(cook_report(
-            cook_id.to_string(),
-            "gate_failed",
-            vec![attempt],
-            None,
-            Some("adopted candidate did not pass the original deterministic gates".to_string()),
-            1,
-        ));
-    }
-    if options.no_finalize {
-        agent_task_lifecycle::finish_candidate_adoption(&record.run_id, None)?;
-        return Ok(cook_report(
-            cook_id.to_string(),
-            "green_no_finalize",
-            vec![attempt],
-            None,
-            Some(
-                "adopted candidate passed deterministic gates; recipe skips finalization"
-                    .to_string(),
-            ),
-            0,
-        ));
-    }
-    agent_task_lifecycle::checkpoint_candidate_adoption(
-        &record.run_id,
-        "finalization",
-        "finalize pull request",
-    )?;
-    let finalization = match finalize_or_load_cook_pr_with_backend(
-        &options,
-        &record.run_id,
-        &promotion,
-        backend,
-    ) {
-        Ok(finalization) => finalization,
-        Err(error) => {
-            agent_task_lifecycle::finish_candidate_adoption(
-                &record.run_id,
-                Some(error.message.clone()),
-            )?;
-            return Err(error);
-        }
-    };
-    agent_task_lifecycle::finish_candidate_adoption(&record.run_id, None)?;
-    let status = finalization["status"]
-        .as_str()
-        .unwrap_or("unknown")
-        .to_string();
-    let exit_code = if status == "review_ready" { 0 } else { 1 };
-    Ok(cook_report(
-        cook_id.to_string(),
-        &status,
-        vec![attempt],
-        Some(finalization),
-        None,
-        exit_code,
-    ))
-}
-
-/// Candidate adoption can prove that an otherwise-red broad command is
-/// inherited only by executing the identical command at the recorded base.
-/// A changed failure fingerprint is deliberately still a hard gate failure.
-fn compare_adoption_gate_failures_to_base(
-    promotion: &mut AgentTaskPromotionReport,
-    source_worktree: &std::path::Path,
-    task_base_sha: &str,
-    run_id: &str,
-) -> Result<()> {
-    if !promotion.status.gate_failed() {
-        return Ok(());
-    }
-    let baseline_root = tempfile::tempdir().map_err(|error| {
-        Error::internal_io(
-            error.to_string(),
-            Some("create candidate-adoption gate baseline".to_string()),
-        )
-    })?;
-    let baseline_path = baseline_root.path().join("base");
-    let output = std::process::Command::new("git")
-        .args(["worktree", "add", "--detach"])
-        .arg(&baseline_path)
-        .arg(task_base_sha)
-        .current_dir(source_worktree)
-        .output()
-        .map_err(|error| {
-            Error::internal_io(
-                error.to_string(),
-                Some("materialize candidate-adoption gate baseline".to_string()),
-            )
-        })?;
-    if !output.status.success() {
-        return Err(Error::internal_io(
-            String::from_utf8_lossy(&output.stderr).trim().to_string(),
-            Some("materialize immutable candidate-adoption gate baseline".to_string()),
-        ));
-    }
-    let baseline_result = (|| -> Result<bool> {
-        // Match promotion's dependency and isolated-runtime setup before the
-        // base command is allowed to serve as comparison evidence.
-        homeboy_core::hygiene::materialize_worktree_dependencies(&baseline_path)?;
-        let mut all_failures_inherited = true;
-        let failed_gate_count = promotion
-            .deterministic_gates
-            .iter()
-            .filter(|gate| gate.status == AgentTaskGateStatus::Failed)
-            .count();
-        if failed_gate_count == 0 {
-            return Ok(false);
-        }
-        let mut compared = 0;
-        for (index, gate) in promotion.deterministic_gates.iter_mut().enumerate() {
-            if gate.status != AgentTaskGateStatus::Failed {
-                continue;
-            }
-            compared += 1;
-            agent_task_lifecycle::checkpoint_candidate_adoption(
-                run_id,
-                "baseline_verification",
-                &format!("baseline gate {compared}/{failed_gate_count}"),
-            )?;
-            let command = gate.command.last().cloned().unwrap_or_default();
-            let baseline_run_dir = homeboy_core::engine::run_dir::RunDir::create()?;
-            let baseline_runtime = homeboy_core::engine::invocation::InvocationGuard::acquire(
-                &baseline_run_dir,
-                &homeboy_core::engine::invocation::InvocationRequirements::default(),
-            )?;
-            let baseline = run_gate_command_with_timeout(
-                &baseline_path,
-                index + 1,
-                &command,
-                gate.visibility,
-                gate.reveal_policy,
-                &baseline_runtime.context().tmp_dir,
-                std::time::Duration::from_secs(5 * 60),
-            )?;
-            let candidate_fingerprint = failure_fingerprint(&gate.stdout, &gate.stderr);
-            let baseline_fingerprint = failure_fingerprint(&baseline.stdout, &baseline.stderr);
-            let matches = baseline.status == AgentTaskGateStatus::Failed
-                && candidate_fingerprint == baseline_fingerprint;
-            gate.baseline_comparison = Some(AgentTaskGateBaselineComparison {
-                base_ref: task_base_sha.to_string(),
-                exit_code: baseline.exit_code,
-                failure_fingerprint: baseline_fingerprint,
-                matches_candidate_failure: matches,
-            });
-            all_failures_inherited &= matches;
-            if matches {
-                gate.accept_inherited_failure();
-            }
-        }
-        Ok(all_failures_inherited)
-    })();
-    let cleanup = std::process::Command::new("git")
-        .args(["worktree", "remove", "--force"])
-        .arg(&baseline_path)
-        .current_dir(source_worktree)
-        .status()
-        .map_err(|error| {
-            Error::internal_io(
-                error.to_string(),
-                Some("remove candidate-adoption gate baseline".to_string()),
-            )
-        })?;
-    if !cleanup.success() {
-        return Err(Error::internal_io(
-            "git worktree remove failed".to_string(),
-            Some("remove candidate-adoption gate baseline".to_string()),
-        ));
-    }
-    let all_failures_inherited = baseline_result?;
-    if all_failures_inherited {
-        promotion.status = crate::agent_task_promotion::AgentTaskPromotionStatus::Applied;
-        for result in &mut promotion.gate_results {
-            if result.status == homeboy_core::gate::HomeboyGateStatus::Failed {
-                result.status = homeboy_core::gate::HomeboyGateStatus::Passed;
-                result.summary = "candidate failure matches the immutable baseline; no candidate regression detected".to_string();
-                result.retryable = Some(false);
-            }
-        }
-    }
-    Ok(())
-}
-
-fn candidate_adoption_source(
-    record: &agent_task_lifecycle::AgentTaskRunRecord,
-    source_request: &AgentTaskRequest,
-) -> Result<(String, Option<PathBuf>, Option<Value>)> {
-    // The authenticated recovery marker is authoritative over a canonical
-    // pre-execution aggregate, which exists only to record the transport error.
-    if let Some(outcome) =
-        agent_task_lifecycle::candidate_adoption_recovery_outcome(record, source_request)
-    {
-        let recovery = outcome.metadata["candidate_adoption_recovery"].clone();
-        return Ok((
-            serde_json::to_string(&outcome).map_err(|error| {
-                Error::internal_json(
-                    error.to_string(),
-                    Some("serialize candidate adoption recovery outcome".to_string()),
-                )
-            })?,
-            None,
-            Some(recovery),
-        ));
-    }
-    if let Ok((source, path)) = agent_task_lifecycle::aggregate_source(&record.run_id) {
-        return Ok((source, Some(path), None));
-    }
-    let (source, path) = promotion_source(&record.run_id)?;
-    Ok((source, path, None))
-}
-
-fn concrete_adoption_ai_model(value: &str) -> Result<String> {
-    let normalized = value.trim();
-    if normalized.is_empty()
-        || value != normalized
-        || value.chars().any(char::is_control)
-        || matches!(
-            normalized.to_ascii_lowercase().as_str(),
-            "not recorded"
-                | "unknown"
-                | "ai-assisted"
-                | "ai assisted"
-                | "legacy caller did not record a model"
-        )
-    {
-        return Err(Error::validation_invalid_argument(
-            "ai_model",
-            "candidate adoption requires a concrete model identifier",
-            None,
-            None,
-        ));
-    }
-    Ok(normalized.to_string())
-}
-
-/// Resolve an existing run first, then recover a deterministic persisted
-/// attempt when a controller stopped after writing its recipe and before
-/// writing the run.
-fn resolve_adoption_target(
-    cook_or_run_id: &str,
-) -> Result<(
-    agent_task_lifecycle::AgentTaskRunRecord,
-    super::AgentTaskCookRecipe,
-)> {
-    if agent_task_lifecycle::run_record_exists(cook_or_run_id)? {
-        let record = agent_task_lifecycle::status(cook_or_run_id)?;
-        let cook_id = record
-            .metadata
-            .get("cook_id")
-            .and_then(Value::as_str)
-            .unwrap_or(cook_or_run_id)
-            .to_string();
-        return Ok((record, super::load_recipe(&cook_id)?));
-    }
-
-    let recipe = if super::recipe_exists(cook_or_run_id)? {
-        super::load_recipe(cook_or_run_id)?
-    } else if let Some(recipe) = super::load_recipe_for_attempt(cook_or_run_id)? {
-        recipe
-    } else {
-        return Err(Error::validation_invalid_argument(
-            "run_or_cook_id",
-            "unknown agent-task run or durable cook id",
-            Some(cook_or_run_id.to_string()),
-            None,
-        ));
-    };
-    let attempt = if recipe.cook_id == cook_or_run_id {
-        resolve_cook_adoption_attempt(&recipe)?
-    } else {
-        recipe
-            .attempts
-            .iter()
-            .find(|attempt| attempt.run_id == cook_or_run_id)
-            .expect("attempt lookup returns a recipe that declares the run id")
-    };
-    if agent_task_lifecycle::run_record_exists(&attempt.run_id)? {
-        return Ok((agent_task_lifecycle::status(&attempt.run_id)?, recipe));
-    }
-
-    agent_task_lifecycle::submit_plan(&attempt.plan, Some(&attempt.run_id))?;
-    agent_task_lifecycle::record_cook_attempt(&recipe.cook_id, attempt.attempt, &attempt.run_id)?;
-    let recovery = Error::internal_unexpected(
-        "recovered orphaned durable cook recipe before provider dispatch".to_string(),
-    );
-    let record = agent_task_lifecycle::record_pre_execution_failure(
-        &attempt.run_id,
-        &attempt.plan,
-        "transport_dispatcher_prepare",
-        &recovery,
-    )?;
-    Ok((record, recipe))
-}
-
-/// A retried cook may have several lifecycle attempts for the same immutable
-/// plan. The earliest is the stable target; different plans require an explicit
-/// run ID so a candidate is never attached to the wrong policy.
-fn resolve_cook_adoption_attempt(
-    recipe: &super::AgentTaskCookRecipe,
-) -> Result<&super::AgentTaskCookRecipeAttempt> {
-    let first = recipe
-        .attempts
-        .first()
-        .expect("loaded cook recipes always contain an attempt");
-    if recipe
-        .attempts
-        .iter()
-        .all(|attempt| attempt.plan == first.plan)
-    {
-        return Ok(first);
-    }
-
-    let attempts = recipe
-        .attempts
-        .iter()
-        .map(|attempt| format!("attempt {}: {}", attempt.attempt, attempt.run_id))
-        .collect::<Vec<_>>()
-        .join(", ");
-    Err(Error::validation_invalid_argument(
-        "cook_recipe.attempts",
-        format!(
-            "candidate adoption by cook id is ambiguous because durable attempt plans differ ({attempts}); rerun with the exact owning run id, for example `homeboy agent-task adopt {}`",
-            first.run_id
-        ),
-        Some(recipe.cook_id.clone()),
-        Some(vec![
-            "Pass an attempt run id to select the candidate's exact recorded policy."
-                .to_string(),
-        ]),
-    ))
 }
 
 /// A bounded collection of independently durable cooks. Each cook retains the
@@ -1618,6 +978,10 @@ fn terminal_executor_identity(
 
 #[cfg(test)]
 mod tests {
+    use super::super::cook_adoption::{
+        adopt_cook_candidate, adopt_cook_candidate_with_dispatcher_and_backend,
+        candidate_adoption_source, concrete_adoption_ai_model, resolve_adoption_target,
+    };
     use super::super::cook_baseline::git_output;
     use super::super::cook_promotion::{
         finalize_cook_pr_with_backend, persisted_promotion_for_attempt,
@@ -1628,6 +992,7 @@ mod tests {
     };
     use crate::agent_task_finalization::{
         AgentTaskPrDurableGateProof, AgentTaskPrFinalizationBackend, AgentTaskPrRef,
+        RealAgentTaskPrFinalizationBackend,
     };
     use crate::agent_task_scheduler::AgentTaskState;
     use homeboy_core::run_lifecycle_record::{
